gradientICA(ica4fts)	R Documentation

Calculates the gradient for the objectiveICA function

Description

Uses a series of C functions to calculate the derivative of the objectiveICA function. Offers a choice between the Huber substitution function or the logCosh function.

Usage


gradientICA(T, E, N, C, PH, method = c("Huber", "Cosh"))

Arguments

T 	 Vector of angles of length p

E 	 Financial time series data of dimension nxd

N 	 Vector of lags that MUST be arranged in ascending order

C 	 Real number C which is used for Huber substitution or Logcosh substitution

PH 	 Phi matrix of weights of dimension q*q

method 	 Choice between use of Huber or logCosh substitution methods

Details

More details to help above

Value

Gradient of objective value

Author(s)

Erjie Ang ea75@cornell.edu

See Also

See

Examples


##
d = 2
n = 10000
p = d*(d-1)/2
E = matrix(rnorm(d*n),n,d)
T = as.matrix(rep(0,p))
N = as.matrix(c(1,2))
q = 2*2*p
C = 1
PH = matrix(rnorm(q*q),q,q)
HCans = gradientICA(T,E,N,C,PH,method = "Huber")
LSans = gradientICA(T,E,N,C,PH,method = "Cosh")

[Package ica4fts version 0.0 Index]